server/ETL/requirements.txt
beautifulsoup4==4.10.0
black==22.1.0
bme680==1.1.1
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
idna==3.3
install==1.3.5
mypy-extensions==0.4.3
mysql-connector-python==8.0.28
pathspec==0.9.0
pigpio==1.78
pihome==0.1
Pillow==9.0.1
platformdirs==2.5.0
protobuf==3.19.4
requests==2.27.1
RPi.GPIO==0.7.1
soupsieve==2.3.1
spidev==3.5
tomli==2.0.1
typing-extensions==4.1.1
urllib3==1.26.8
